% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{giff_data}
\alias{giff_data}
\title{Returns the dataset of a given set and variant}
\usage{
giff_data(.set, .variant = NULL)
}
\arguments{
\item{.variant}{}
}
\value{

}
\description{
Returns the dataset of a given set and variant
}
